Declare a code dependency in the code instead of in the data
When using the test function persistent_key_load_key_from_storage with
DERIVE_KEY, there's a dependency on HKDF-SHA-256. Since this
dependency is in the code, declare it there rather than with the data.
Also, if the dependency is not met, do what can be done, and check
that the step that's expected to fail does fail as expected.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>