Remove one more obsolete asymmetric key import test case
A long time ago, the PSA key import format included metadata in the
key material which was redundant with the key type conveyed as a key
attribute. The test suite therefore had test cases validating that an
attempt to import a key with different key types conveyed in the key
material and in the attributes was properly rejected. Now that the
format of the key material no longer includes type information, these
test cases are just generic wrong-size tests. Remove one more that
just broke due to some cases returning NOT_SUPPORTED instead of
INVALID_ARGUMENT.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>