Detect and report mutex usage errors (+FIXME move test_info)
If the mutex usage verification framework is enabled and it detects a
mutex usage error, report this error and mark the test as failed.

This detects most usage errors, but not all cases of using
uninitialized memory (which is impossible in full generality) and not
leaks due to missing free (which will be handled in a subsequent commit).

FIXME: This commit also moves test_info around. This should be
resolved by rebasing on top of
https://github.com/ARMmbed/mbedtls/pull/4054 first.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>