Count and report non-freed mutexes
If a call to mbedtls_mutex_init() is missing the corresponding call to
mbedtls_mutex_free(), a memory leak detector will report it as a leak
at the end of the test suite. To make the leak easier to locate,
print a message after the test case where it happens.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>