Split test storage into save and read tests
Save tests are for forward compatibility: save a key in the current
format and check that it has the expected format so that future
versions will still be able to read it.

Read tests are for backward compatibility: read a key in the format of
a past version and check that this version can use it. For read tests,
exercise the key unless it is meant to test metadata storage only.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>