PKCS#12/MD2: use de facto standard block size
PKCS#12 explicitly specifies that v = 512 bits = 64 bytes for MD2,
but this is an incorrect application of the general rule of using the
size of "the message input to the compression function", i.e. the
block size, which is 128 bits = 16 bytes.

Mbed TLS classically used the explicit value from the specification,
but this made it incompatible with the rest of the world, so change to
follow the generic rule from the specification, which is also the de
facto standard.

Fix #4267.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>