Use a wrapper macro when reading MPIs from test data
Replace calls to mbedtls_mpi_read_string() with a wrapper read_mpi() when
reading test data except for the purpose of testing
mbedtls_mpi_read_string() itself. The wrapper lets the test data control
precisely how many limbs the constructed MPI has.

This will make the tests easier to maintain if the behavior of the library
function deviates from the exact behavior desired for testing with respect
to the size of the result. Having this level of indirection also makes it
easier to backport tests to branches where the library function behaves
differently.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>