Make socket error checks more robust on Windows
According to
https://docs.microsoft.com/en-us/windows/win32/winsock/socket-data-type-2 :
> Windows Sockets handles have no restrictions, other than that the
> value INVALID_SOCKET is not a valid socket.

Use the symbolic name INVALID_SOCKET instead of the constant -1 for an
invalid socket handle on all platforms. Replace `fd < 0` check by `fd ==
INVALID_SOCKET`.

This is unlikely to cause a difference in practice according to
https://stackoverflow.com/questions/1953639/is-it-safe-to-cast-socket-to-int-under-win64
but it makes the code more robust for 32-bit Windows. (For 64-bit Windows,
the `fd` field of `mbedtls_net_context` needs to have a different type.)

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>