Make constant-flow functions non-static
This is in preparation for moving them to a module of their own.

There were three functions called mbedtls_cf_size_bool_eq(): one in
base64.c, one in bignum.c and one in ssl_msg.c. They had exactly the same
behavior, except that the one in base64.c returned unsigned char and the
other two returned size_t. Keep only the one in bignum.c and remove the
other two. Since the function returns 0 or 1 and is not used in a context
where the width of the type would matter, the size_t variant can be used in
place of the unsigned char variant.

This temporarily breaks the build in configurations with base64 or ssl but
not bignum. This temporarily breaks the build with compilers that insist
that non-static functions must have been declared before their definition.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>