Zeroize local MAC variables
Zeroize local MAC variables used for CBC+HMAC cipher suites. In encryption,
this is just good hygiene but probably not needed for security since the
data protected by the MAC that could leak is about to be transmitted anyway.
In decryption, this could be a security issue since an adversary could learn
the MAC of data that they were trying to inject.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>