Add test code for PSA size macros that should be constant
PSA macros are supposed to expand to constant expressions when their
arguments are constant expressions. Verify this for the macros that are most
at risk of accidentally including non-constant parts: buffer size
calculation macros that have arguments. This will break the build if a
tested macro is non-constant when building with `gcc -Wvla`.

For this commit, I manually extracted the list of macros to test from
crypto_sizes.h. The command I used is in a comment.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>