Fix non-constant size macros
Fix the implementation of some PSA buffer size macros that expanded to
non-constant expressions. Fixes #4320.

All the non-constant expressions were of the form `((void)(param), 0)`
to convey that `param` is deliberately unused (at least in some builds).
Replace this with `0 & (param)`, which is a constant expression and which
typical compilers don't seem to complain about.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>