WIP Test pk_parse_public_key with truncated data or trailing garbage
Test all non-empty truncations, test the exact data for reference, test
extra nulls at the end of PEM, and test DER with appended garbage.

PEM parsing ignores trailing garbage, so test with trailing garbage as well.
THIS IS A WORK IN PROGRESS and that part of the tests is failing.

Also check the expected error code when a feature (algorithm or curve) is
not included in the build.

Test data in this commit: tests/data_files/rsa_pkcs8_1024_public.der and
tests/data_files/ec_256_pub.pem converted to the requisite input format.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>