Support testing against openssl/gnutls server in RSA-only build
When testing against a foreign server, we always use a specific key and
certificate. Arrange to pick either ECDSA or RSA depending on what is
available in the Mbed TLS build-time configuration.

Pick ECDSA if possible, both because it's faster than RSA and because it was
the original choice and some test cases might not work with RSA. But do fall
back to RSA so we can test in a configuration without ECC. These parameters
are ignored when using a pre-shared key.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>