When using a specific private key, require it to be supported
When running a test that specifies a specific key file to use with s_server2
or s_client2, require the key's algorithm to be supported at compile time.

This is only implemented for a small subset of known keys, enough to pass
the CI at this point in time. The detection is based on hard-coding
knowledge about our test keys. The detection is only valid for
signature-based or RSA-decryption-based ciphersuites, ignoring the
possibility of e.g. a static ECDH key, or a configuration with RSA but not
PKCS#1v1.5.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>