Don't force a ciphersuite into a version that isn't supported
When ssl_client2 is run with both force_ciphersuite=... and a range of
protocol versions restricted at runtime (min_version=... or
max_version=...), it adjusts the range of protocol versions to versions
supported by the ciphersuite. This was done incorrectly in the case where
some protocol versions are also disabled at compile time, which made it
offer the ciphersuite with a version that it didn't support.

To reproduce: in config-mini-tls1_1.h,

    ssl-opt.sh -f 'keyUsage cli: DigitalSignature+KeyEncipherment, RSA: OK'
    ssl-opt.sh -f 'keyUsage cli: KeyEncipherment, RSA: OK'

Cap the protocol versions to the supported range.

Signed-off-by: Gilles Peskine <Gilles.Peskine@arm.com>